+++ client-sdk/ts-web/ext-utils/sample-page/src/index.js
@@ -1,6 +1,7 @@
 // @ts-check
 
 import * as oasis from '@oasisprotocol/client';
+import * as oasisRT from '@oasisprotocol/client-rt';
 
 import * as oasisExt from '../..';
 
@@ -34,10 +35,39 @@ export const playground = (async function () {
         oasis.staking.addressToBech32(await oasis.staking.addressFromPublicKey(publicKey)),
     );
 
+    const dst = oasis.signature.NaclSigner.fromRandom('this key is not important');
+    const tw = oasis.staking
+        .transferWrapper()
+        .setNonce(101n)
+        .setFeeAmount(oasis.quantity.fromBigInt(102n))
+        .setFeeGas(103n)
+        .setBody({
+            to: await oasis.staking.addressFromPublicKey(dst.public()),
+            amount: oasis.quantity.fromBigInt(104n),
+        });
     console.log('requesting signature');
-    const signature = await signer.sign('invalid/sample-message: v0', new Uint8Array([1, 2, 3]));
+    await tw.sign(signer, 'fake-chain-context-for-testing');
     console.log('got signature');
-    console.log('signature base64', toBase64(signature));
+    console.log('signature base64', toBase64(tw.signedTransaction.signature.signature));
+
+    const rtw = new oasisRT.accounts.Wrapper(oasis.misc.fromString('fake-runtime-id-for-testing'))
+        .callTransfer()
+        .setBody({
+            to: await oasis.staking.addressFromPublicKey(dst.public()),
+            amount: [oasis.quantity.fromBigInt(105n), oasis.misc.fromString('TEST')],
+        })
+        .setSignerInfo([
+            {
+                address_spec: {signature: {ed25519: publicKey}},
+                nonce: 106n,
+            },
+        ])
+        .setFeeAmount([oasis.quantity.fromBigInt(107n), oasisRT.token.NATIVE_DENOMINATION])
+        .setFeeGas(108n);
+    console.log('requesting signature');
+    await rtw.sign([signer], 'fake-chain-context-for-testing');
+    console.log('got signature');
+    console.log('signature base64', toBase64(rtw.unverifiedTransaction[1][0].signature));
 })();
 
 playground.catch((e) => {