+++ client-sdk/ts-web/rt/docs/changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Unreleased changes
+
+New features:
+
+- `event.Visitor.visit` now returns whether it had a handler for the given
+  event.
+- Added `transaction.SignatureMessageHandlersWithChainContext` for use with
+  `oasis.signature.visitMessage`.
+- Added `transaction.visitCall` for the case where that message is a runtime
+  transaction.
+
 ## v0.2.0-alpha2
 
 Spotlight change: